[issue 203] add validations.py and validation addr checksum throughout
