+++ tests/eth-module/test_iban.py
@@ -38,9 +38,23 @@ def test_createIndirect(value, expected, web3):
             '0x0000a5327eab78357cbf2ae8f3d49fd9d90c7d22',
             'XE0600DQK33XDTYUCRI0KYM5ELAKXDWWF6',
         ),
+        (
+            '0x606060405261022e806100126000396000f360606040523615610074576000357c01000000000000000000000000000000000000000000000000000000009004806316216f391461007657806361bc221a146100995780637cf5dab0146100bc578063a5f3c23b146100e8578063d09de08a1461011d578063dcf537b11461014057610074565b005b610083600480505061016c565b6040518082815260200191505060405180910390f35b6100a6600480505061017f565b6040518082815260200191505060405180910390f35b6100d26004808035906020019091905050610188565b6040518082815260200191505060405180910390f35b61010760048080359060200190919080359060200190919050506101ea565b6040518082815260200191505060405180910390f35b61012a6004805050610201565b6040518082815260200191505060405180910390f35b6101566004808035906020019091905050610217565b6040518082815260200191505060405180910390f35b6000600d9050805080905061017c565b90565b60006000505481565b6000816000600082828250540192505081905550600060005054905080507f3496c3ede4ec3ab3686712aa1c238593ea6a42df83f98a5ec7df9834cfa577c5816040518082815260200191505060405180910390a18090506101e5565b919050565b6000818301905080508090506101fb565b92915050565b600061020d6001610188565b9050610214565b90565b60006007820290508050809050610229565b91905056',
+            TypeError,
+        ),
+        (
+            '0xd3CDA913deB6f67967B99D67aCDFa1712C293601',
+            ValueError,
+        ),
     ),
 )
 def test_fromAddress(value, expected, web3):
+
+    if isinstance(expected, type) and issubclass(expected, Exception):
+        with pytest.raises(expected):
+            web3.eth.iban.fromAddress(value).toString()
+        return
+
     actual = web3.eth.iban.fromAddress(value).toString()
     assert actual == expected
